feat: pss-imp-def-03r-compilation-002 - Move loading, loadedsource, and runtimeconfig behind compilation
