+++ docs/internal/processes/runtime-cleanup-relevant-files.md
@@ -774,6 +774,13 @@ imports avoid Wire/Runtime/Petri/peer/sibling-lease paths and public
 loading/loadedsource/runtimeconfig packages, and `wire/wire.go` constructs from
 injected ports without selecting Runtime/Petri implementations or sibling
 catalog/authoring_layout/validation/snapshots_portability/distribution leases.
+Compilation-owned loading/loadedsource/runtimeconfig implementation lives under
+`internal/services/compilation/{loading,loadedsource,runtimeconfig}`; public
+`loading/`, `loadedsource/`, and `runtimeconfig/` remain transitional re-exports
+for `pkg/wire` and in-owner callers until peer imports retire in later stories.
+Factory Definitions `wire/wire.go` composes the compilation subservice from the
+nested loader ports and delegates `CompileEffectiveFactorySource` on the returned
+root `Service`.
 Snapshot
 slices stay on the singular `Service` via `CaptureFactorySnapshot`,
 `PrepareFactorySnapshotImport`, and `MaterializeFactorySnapshot` returning
